Room generation
there is now sufficient free space inbetween the walls but the binary tree is not complete so there are no other rooms yet. I considered a portal between levels like zelda where the stair case leads to the next level but i finaly settled on the corridor where it is a safe space from the monsters. The algorithm is no longer hunt and kill because the gaming experience would be sub par as there is no room to evaid an avoid the mobs