Roooms are now carved
this was so much work - so the main problem initially was that the rooms were being carved either too close together or too far apart or in the worst case all the rooms were overlapping. to solve this in the horizontal and vertical split function i incremented the min col and nim row  - conditionally to account for the lack in each one when compared to the self.MAX. when actually carving the rooms  we dont want every room to be filled resulting in a uniform dungon. therefore if the room was determined to be above 80 percent filled then we would reduce the width and height of said leaf accordingly. now the section assigned to the leaf is defined its time to make the room as we dont want tiny rooms making the game dificult to play i set it to within 60 percent fo the available section