preperation for scrolling
first removed floor group as there were no usages. then moved the rendering of the background to inside the game loop so that the map will update each cycle. then instead of adding to an all sprite group i created a background sprite groups for the wall and floors and any other furture items. this means that you can flip from one frame to the next without an overlap of data. using old wall and walls. seperate foreground sprite group for player so that when drawing it is not covered by the floor and so can add any monsters or npcs in the future.