simplify and fix locatew
initially locate funciton only worked when the portal was generated withing the detection zone

once the locate funciton searched the whole map then it was able to smoothly located the portal.

as the map is large the initial scroll speed is very high but once gameplay begins it become choppy. therefore by flagging when the player is in the detections zone and changeing the scroll speed the experience is much more enjoyable